refactor(device-hub): split usbip.rs into a module directory
The single file had grown past 1300 lines across the transport commands,
status probing, elevated IPC, VHCI residual cleanup, and output parsing.
Splits it into src-tauri/src/usbip/ with one file per responsibility:

- mod.rs: shared types, constants, input validation, re-exports
- commands.rs: the five tauri commands
- status.rs: transport status probing
- exec.rs: locating and running usbip.exe plus output parsing
- device.rs: VHCI residual detection and cleanup
- elevated.rs: the elevated named-pipe IPC and helper entry point

Pure code move: public surface, command registration, cfg gates, and all
ten tests are unchanged; internal types stay module-private (private items
remain visible to the child modules). Verified with a three-lens
adversarial audit (completeness, body drift, wiring) over the old and new
sources: 61 items on both sides, zero body drift after normalizing the
path and visibility adjustments, cargo test 168 passed.
